requirements.txt
click==7.1.2
python-dotenv==0.14.0
mimesis==4.1.2
pytest==6.1.1